ALPHA 3.2: TOOLS
Implemented WoodenShovel, WoodenPickaxe, WoodenAxe
Finalized Tool and ToolItem structure

+++ AllegroGame/AirTile.h
@@ -11,7 +11,7 @@ public:
     virtual std::string GetName() const final;
 
     virtual ToolType GetOptimalToolType() const final;
-    virtual int GetMiningResistance() const final;
+    virtual float GetMiningResistance() const final;
 
     virtual Tile* Clone(World* w, int x, int y, Direction d) const final;
