+++ AllegroGame/BasicConveyorTile.h
@@ -36,7 +36,7 @@ public:
     virtual bool CanWalkThrough() const final;
 
     virtual ToolType GetOptimalToolType() const final;
-    virtual int GetMiningResistance() const final;
+    virtual float GetMiningResistance() const final;
     virtual std::string GetName() const final;
 
     virtual bool DoesTickUpdates() const final;
